% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompPASSplus.R
\name{CompPASSplus}
\alias{CompPASSplus}
\title{CompPASSplus}
\usage{
CompPASSplus(datInput)
}
\arguments{
\item{datInput}{A dataframe with column names: idRun, idBait, idPrey, 
countPrey. Each row represent one unique protein captured in one 
pull-down experiment}
}
\value{
A data frame consists of unique bait-prey pairs with 
Z-score, S-score,D-score and WD-score indicating interacting probabilities.
}
\description{
CompPASSplus
In addition to CompPASS, this function computes entropy and normalized 
WD-score. interacting proteins from non-specific background and false 
positive identifications on the basis of CompPASS scores. The source 
code for this function was based on the source code. 
 \url{https://github.com/dnusinow/cRomppass}
}
\examples{
data(TestDatInput)
CompPASSplus(TestDatInput)
}
\references{
Huttlin, Edward L., et al. "The BioPlex network: a systematic 
exploration of the human interactome." Cell 162.2 (2015): 425-440. 
\url{https://doi.org/10.1016/j.cell.2015.06.043}

Huttlin, Edward L., et al. "Architecture of the human 
interactome defines protein communities and disease networks." 
Nature 545.7655 (2017): 505. 
\url{https://www.nature.com/articles/nature22366}
}
\author{
Qingzhou Zhang, \email{zqzneptune@hotmail.com}
}
