% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DICE.R
\name{DICE}
\alias{DICE}
\title{DICE}
\usage{
DICE(datInput)
}
\arguments{
\item{datInput}{A dataframe with column names: idRun, idPrey.
Each row represent one unique protein captured in one pull-down 
experiment.}
}
\value{
A dataframe consists of pairwise combindation of preys identified in 
the input with DICE scores.
}
\description{
DICE
The Dice coefficient is used to score the interaction affinity between two proteins.
}
\examples{
data(TestDatInput)
datScore <- DICE(TestDatInput)
head(datScore)
}
\references{
Bing Zhang et al., From pull-down data to protein interaction 
networks and complexes with biological relevance, Bioinformatics, 
Volume 24, Issue 7, 1 April 2008, Pages 979–986, 
\url{https://doi.org/10.1093/bioinformatics/btn036}
}
\author{
Qingzhou Zhang, \email{zqzneptune@hotmail.com}
}
