% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetPPN}
\alias{GetPPN}
\title{GetPPN}
\usage{
GetPPN(mat)
}
\arguments{
\item{mat}{A matrix of peptide/spectra count in the format of 
proteins (column) by AP-MS runs (row)}
}
\value{
A matrix of proteins by proteins indicating sum of minimum 
peptide/spectra count across all AP-MS runs.
}
\description{
This function returns sum of minimum peptide/spectra count across AP-MS
runs.
}
\examples{
mat <- matrix(c(3, 4, 5, 3, 4, 2), byrow = TRUE, ncol = 2)
GetPPN(mat)

}
\references{
Guruharsha, K. G., et al. "A protein complex network of 
Drosophila melanogaster." Cell 147.3 (2011): 690-703.
\url{https://doi.org/10.1016/j.cell.2011.08.047}
}
\author{
Qingzhou Zhang, \email{zqzneptune@hotmail.com}
}
