% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestDatInput.R
\docType{data}
\name{TestDatInput}
\alias{TestDatInput}
\title{Test data for SMAD}
\format{A data frame with 5000 rows and 5 variables}
\usage{
data(TestDatInput)
}
\description{
It is a subset of unfiltered BioPlex 2.0 consisting of
apoptosis as bait proteins
}
\details{
\itemize{
  \item idRun A unique identifier for one AP-MS run
  \item idBait A unique identifier for the bait protein
  \item idPrey A unique identifier for the prey protein
  \item countPrey Sepctra/Peptider count for the prey protein
  \item lenPrey Protein length for the prey protein
}
}
\keyword{datasets}
