% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PE.R
\name{PE}
\alias{PE}
\title{PE}
\usage{
PE(datInput, rBait = 0.37, cntPseudo = 1)
}
\arguments{
\item{datInput}{A dataframe with column names: idRun, idBait, idPrey, 
Each row represent one unique protein captured in one pull-down 
experiment.}

\item{rBait}{The value of the 'r' parameter as desribed in the publication.}

\item{cntPseudo}{The value of the 'pseudo count' parameter.}
}
\value{
A dataframe consists of protein-protein interactions from both Spoke 
and Matrix model.
}
\description{
PE
Incorporated both spoke and matrix model.
}
\examples{
data(TestDatInput)
datScore <- PE(TestDatInput, 0.37, 1)
head(datScore)
}
\references{
Collins, Sean R., et al. "Toward a comprehensive atlas of the 
physical interactome of Saccharomyces cerevisiae." 
Molecular & Cellular Proteomics 6.3 (2007): 439-450.
\url{https://doi.org/10.1074/mcp.M600381-MCP200}
}
\author{
Qingzhou Zhang, \email{zqzneptune@hotmail.com}
}
